- (Diep) - Refactored more top level files into the engine and subsystem folder.
