+++ src/app/diep/engine/subsystems/diep.enemy-spawner.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
-import { Enemy, EnemyType, EnemySpawnWeight } from './diep.interfaces';
-import { EnemyRegistry } from './enemies/enemy.registry';
+import { Enemy, EnemyType, EnemySpawnWeight } from '../../diep.interfaces';
+import { EnemyRegistry } from '../../enemies/enemy.registry';
 
 const ENEMY_SPAWN_WEIGHTS: EnemySpawnWeight[] = [
     { type: 'SMASHER', weight: 0.05 },
