+++ src/app/diep/engine/subsystems/diep.projectile.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Bullet, TrailSegment, Player } from '../diep.interfaces';
+import { Bullet, TrailSegment, Player } from '../../diep.interfaces';
 
 @Injectable({
     providedIn: 'root'
